requirements.txt
google-cloud-datastore
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
PyJWT==1.7.1
pytz==2018.9
six==1.12.0
Werkzeug==0.15.2
